skills/fabius-cohors/references/agents/python/safety-plugins/pyproject.toml
[project]
name = "safety-plugins"
version = "0.1.0"
description = "Safety Plugins for ADK"
authors = [
    { name = "sammalik-lgtm", email = "sammalik@google.com" }
]
license = {text = "Apache License 2.0"}
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "absl-py>=2.3.1",
    "google-adk>=1.7.0",
    "google-auth>=2.30.0",
    "google-cloud-modelarmor>=0.2.8",
    "google-genai>=1.5.0",
    "pdfplumber>=0.11.5",
    "pydantic>=2.11.7",
    "pydantic-core>=2.33.2",
    "pydantic-settings>=2.10.0",
    "python-dotenv>=1.0.1",
    "typing-extensions>=4.14.0",
    "google-cloud-aiplatform[adk,agent-engines]>=1.93.0",
]

[dependency-groups]
dev = [
    "agent-starter-pack>=0.14.1",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "ruff>=0.4.6",
]

[tool.ruff]
# Point back to the Master template in the root
extend = "../../../pyproject.toml"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"safety_plugins/__init__.py" = ["E402", "I001"]

[tool.ruff.lint.isort]
known-first-party = ["safety_plugins"]

[tool.pytest.ini_options]
pythonpath = "."
asyncio_default_fixture_loop_scope = "function"

[tool.agent-starter-pack]
example_question = "Can you run the short sum tool for me?"

[tool.agent-starter-pack.settings]
agent_directory = "safety_plugins"

[tool.hatch.build.targets.wheel]
# Ensure Hatchling finds the safety package even if there's a name collision
packages = ["safety_plugins"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"